refactor(settings): split the settings window around a controller
SettingsView kept a full copy of the model snapshot in state and needed two
merge helpers, two refresh paths, a version counter, and per-command
bookkeeping to keep the copy honest. A SettingsController now owns only what
the window authors locally: the draft profile, the selection, the tab, and
presentation state. Everything else is read from the model. The three
reconciliation rules survive as controller behavior with their tests ported.

Views receive the controller instead of up to eighteen closures, so SwiftUI
can compare their inputs again. The 3400-line file is split by ownership into
formatting, analysis, window chrome, style, and one file per view family.
Closure bindings became key-path projections, the plot geometry is shared by
the grid and the animated curve, card borders use strokeBorder, label rows go
through SettingRow, and the dead helpers found in the audit are gone.
